fix: build error 수정, 스플래시 스크린 처리 위치 변경
