--- frontend/src/components/SplashScreen/index.tsx
@@ -1,5 +1,0 @@
-const SplashScreen = () => {
-  return <div>스플래시 스크린</div>;
-};
-
-export default SplashScreen;
